myproject/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
mysqlclient>=2.1
python-dotenv>=1.0